+++ packages/checks/src/exists-not-null.ts
@@ -1,1 +1,2 @@
-export const existsAndNotNull = (x: any) => x != null;
+export const existsAndNotNull = <T>(x: T | null | undefined): x is T =>
+    x != null;