+++ packages/matrices/src/transpose.ts
@@ -2,7 +2,8 @@ import { setC, setC4 } from "@thi.ng/vectors";
 import { MatOpM } from "./api";
 
 /**
- * Writes transposition of M22 `m` to `out`.
+ * Writes transposition of 2x2 matrix `m` to `out`. Creates new matrix
+ * if `out` is `null`
  *
  * @param out
  * @param m
@@ -11,7 +12,8 @@ export const transpose22: MatOpM = (out, m) =>
     setC4(out || [], m[0], m[2], m[1], m[3]);
 
 /**
- * Writes transposition of M33 `m` to `out`.
+ * Writes transposition of 3x3 matrix `m` to `out`. Creates new matrix
+ * if `out` is `null`
  *
  * @param out
  * @param m
@@ -20,7 +22,8 @@ export const transpose33: MatOpM = (out, m) =>
     setC(out || [], m[0], m[3], m[6], m[1], m[4], m[7], m[2], m[5], m[8]);
 
 /**
- * Writes transposition of M44 `m` to `out`.
+ * Writes transposition of 4x4 matrix `m` to `out`. Creates new matrix
+ * if `out` is `null`
  *
  * @param out
  * @param m