+++ packages/webgl/src/readpixels.ts
@@ -1,0 +1,34 @@
+import { ITexture, ReadableTextureFormat, TextureType } from "./api/texture";
+import { FBO } from "./fbo";
+
+export const readPixels = <
+    T extends Uint8Array | Uint16Array | Uint32Array | Float32Array
+>(
+    gl: WebGLRenderingContext,
+    x: number,
+    y: number,
+    w: number,
+    h: number,
+    format: ReadableTextureFormat,
+    type: TextureType,
+    out: T
+) => {
+    gl.readPixels(x, y, w, h, format, type, out);
+    return out;
+};
+
+export const readTexture = <
+    T extends Uint8Array | Uint16Array | Uint32Array | Float32Array
+>(
+    gl: WebGLRenderingContext,
+    tex: ITexture,
+    format: ReadableTextureFormat,
+    type: TextureType,
+    out: T
+) => {
+    const fbo = new FBO(gl, { tex: [tex] });
+    gl.readPixels(0, 0, tex.size[0], tex.size[1], format, type, out);
+    fbo.unbind();
+    fbo.release();
+    return out;
+};