+++ packages/webgl/src/utils.ts
@@ -5,13 +5,7 @@ import {
     Sym,
     Term
 } from "@thi.ng/shader-ast";
-import { ShaderOpts } from "./api";
-
-export const isGL2Context = (
-    gl: WebGLRenderingContext
-): gl is WebGL2RenderingContext =>
-    typeof WebGL2RenderingContext !== "undefined" &&
-    gl instanceof WebGL2RenderingContext;
+import { ShaderOpts } from "./api/shader";
 
 export const positionAttrib = (
     opts: Partial<ShaderOpts<any>>,